Add PrivilegeController implementation
